Add rack app to use in end-to-end tests
End to end tests should be marked with the "endtoend" tag. This will
start up a Rack app that serves a website with three tiers of
links/pages.

`spec/spec_helper.rb` contains the code to start up the server. It
obtains a random open port from the OS, uses `rackup` to start it and
after the test is finished, kills the server process.