Add crawler
The crawler navigates a website, creating pages, breadth-first.

Breadth-first is necessary because every page may have navigation links
and the crawler does not visit a page twice. If the search was done
depth-first, it would appear as though the first nav link has a child of
the second link, which has a child of the third link and so on.